+++ src/Sulu/Bundle/TagBundle/Admin/TagAdmin.php
@@ -15,7 +15,7 @@ use Sulu\Bundle\AdminBundle\Admin\Admin;
 use Sulu\Bundle\AdminBundle\Admin\Navigation\NavigationItem;
 use Sulu\Bundle\AdminBundle\Admin\Navigation\NavigationItemCollection;
 use Sulu\Bundle\AdminBundle\Admin\View\RouteBuilderFactoryInterface;
-use Sulu\Bundle\AdminBundle\Admin\View\RouteCollection;
+use Sulu\Bundle\AdminBundle\Admin\View\ViewCollection;
 use Sulu\Bundle\AdminBundle\Admin\View\ToolbarAction;
 use Sulu\Component\Security\Authorization\PermissionTypes;
 use Sulu\Component\Security\Authorization\SecurityCheckerInterface;
@@ -59,7 +59,7 @@ class TagAdmin extends Admin
         }
     }
 
-    public function configureViews(RouteCollection $routeCollection): void
+    public function configureViews(ViewCollection $viewCollection): void
     {
         $formToolbarActions = [];
         $listToolbarActions = [];
@@ -82,7 +82,7 @@ class TagAdmin extends Admin
         }
 
         if ($this->securityChecker->hasPermission(static::SECURITY_CONTEXT, PermissionTypes::EDIT)) {
-            $routeCollection->add(
+            $viewCollection->add(
                 $this->routeBuilderFactory->createListRouteBuilder(static::LIST_ROUTE, '/tags')
                     ->setResourceKey('tags')
                     ->setListKey('tags')
@@ -92,12 +92,12 @@ class TagAdmin extends Admin
                     ->setEditRoute(static::EDIT_FORM_ROUTE)
                     ->addToolbarActions($listToolbarActions)
             );
-            $routeCollection->add(
+            $viewCollection->add(
                 $this->routeBuilderFactory->createResourceTabRouteBuilder(static::ADD_FORM_ROUTE, '/tags/add')
                     ->setResourceKey('tags')
                     ->setBackRoute(static::LIST_ROUTE)
             );
-            $routeCollection->add(
+            $viewCollection->add(
                 $this->routeBuilderFactory->createFormRouteBuilder('sulu_tag.add_form.details', '/details')
                     ->setResourceKey('tags')
                     ->setFormKey('tag_details')
@@ -106,13 +106,13 @@ class TagAdmin extends Admin
                     ->addToolbarActions($formToolbarActions)
                     ->setParent(static::ADD_FORM_ROUTE)
             );
-            $routeCollection->add(
+            $viewCollection->add(
                 $this->routeBuilderFactory->createResourceTabRouteBuilder(static::EDIT_FORM_ROUTE, '/tags/:id')
                     ->setResourceKey('tags')
                     ->setBackRoute(static::LIST_ROUTE)
                     ->setTitleProperty('name')
             );
-            $routeCollection->add(
+            $viewCollection->add(
                 $this->routeBuilderFactory->createFormRouteBuilder('sulu_tag.edit_form.details', '/details')
                     ->setResourceKey('tags')
                     ->setFormKey('tag_details')
